depot tab fixed, all reserved signals are green
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@677 8aca7d54-2c30-db11-9de9-000461428c89

+++ besch/obj_besch.h
@@ -16,6 +16,8 @@
 
 #include "../simtypes.h"
 
+#include <stdlib.h>
+
 /*
  *  Autor:
  *      Volker Meyer
@@ -37,12 +39,12 @@ class obj_besch_t {
 		obj_besch_t** node_info;
 
 	public:
-		void* operator new(unsigned int size)
+		void* operator new(size_t size)
 		{
 			return ::operator new(size);
 		}
 
-		void* operator new(unsigned int size, unsigned int extra)
+		void* operator new(size_t size, unsigned extra)
 		{
 			return ::operator new(size + extra);
 		}
